precess Email
This function is used to process email so that all words are separated
individually. And the important parts of this function include three
functions. First, removing some confusing words and punctuation.
Second, using several initial character of words to mark features
instead of exact words. Third, returning words that appear in the
vocabulary list.